Added Parcel as application bundler
