+++ .cache/12/a1f166b2fe89a29b2ab7c70d996d19.json
@@ -1,0 +1,1 @@
+{"id":"src/index.ts","dependencies":[{"name":"/Users/marcustallberg/Documents/workspace/blur/.babelrc","includedInParent":true,"mtime":1554983657000},{"name":"/Users/marcustallberg/Documents/workspace/blur/package.json","includedInParent":true,"mtime":1571827266123}],"generated":{"js":"console.log('hello world');"},"sourceMaps":{"js":{"mappings":[{"source":"src/index.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":1,"column":0}},{"source":"src/index.ts","name":null,"original":{"line":1,"column":7},"generated":{"line":1,"column":7}},{"source":"src/index.ts","name":null,"original":{"line":1,"column":8},"generated":{"line":1,"column":8}},{"source":"src/index.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":1,"column":11}},{"source":"src/index.ts","name":null,"original":{"line":1,"column":12},"generated":{"line":1,"column":12}},{"source":"src/index.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":1,"column":25}}],"sources":{"src/index.ts":"console.log('hello world')"},"lineCount":null}},"error":null,"hash":"cd3b159937414bdd400bef66feee376e","cacheData":{"env":{}}}
