--- README.md
@@ -1,30 +1,0 @@
-# Blur Fi
-
-The best jQuery plugin ever.
-
-## Getting Started
-Download the [production version][min] or the [development version][max].
-
-[min]: https://raw.github.com/marcus/blur_fi/master/dist/blur_fi.min.js
-[max]: https://raw.github.com/marcus/blur_fi/master/dist/blur_fi.js
-
-In your web page:
-
-```html
-<script src="jquery.js"></script>
-<script src="dist/blur_fi.min.js"></script>
-<script>
-jQuery(function($) {
-  $.awesome(); // "awesome"
-});
-</script>
-```
-
-## Documentation
-_(Coming soon)_
-
-## Examples
-_(Coming soon)_
-
-## Release History
-_(Nothing yet)_
