+++ index.html
@@ -2,112 +2,6 @@
 <html lang="en">
 <head>
 	<meta charset=utf-8>
-	<title>blur.fi - Pixelate your images in your browser</title>
-	<meta property="og:title" content="blur.fi" />
-	<meta property="og:description" content="Pixelate your images in your browser." />
-	<meta property="og:type" content="website" />
-	<meta property="og:site_name" content="blur.fi" />
-	<meta property="og:image" content="http://www.blur.fi/images/fbshare.png" />
-	<meta property="og:url" content="http://www.blur.fi" />				
-	<meta property="fb:admins" content="722419471"/>
-	<meta name="author" content="Marcus Tallberg"/>
-	<meta name="description" content="Pixelate your images in your browser." />
-	<meta name="keywords" content="html5 canvas javascript pixelate blur drag&amp;drop" />
-	<link rel="stylesheet" href="css/style-min.css" type="text/css" />
-	<link href='http://fonts.googleapis.com/css?family=Graduate' rel='stylesheet' type='text/css'>
-	<script type="text/javascript" src="js/libs/jquery-1.7.1.min.js"></script>
-	<script type="text/javascript">
-	  var _gaq = _gaq || [];
-	  _gaq.push(['_setAccount', 'UA-28665491-1']);
-	  _gaq.push(['_trackPageview']);
-	  (function() {
-	    var ga = document.createElement('script'); ga.type = 'text/javascript'; ga.async = true;
-	    ga.src = ('https:' == document.location.protocol ? 'https://ssl' : 'http://www') + '.google-analytics.com/ga.js';
-	    var s = document.getElementsByTagName('script')[0]; s.parentNode.insertBefore(ga, s);
-	  })();
-	</script>
-</head>
-<body>
-	<header>
-		<h1>blur.fi</h1>
-		<div class="sharing"><iframe src="//www.facebook.com/plugins/like.php?href=http%3A%2F%2Fwww.blur.fi&amp;send=false&amp;layout=button_count&amp;width=200&amp;show_faces=false&amp;action=like&amp;colorscheme=light&amp;font&amp;height=21&amp;appId=138968016196944" scrolling="no" frameborder="0" style="border:none; overflow:hidden; width:100px; height:21px;margin-bottom:5px;" allowTransparency="true"></iframe><a href="https://twitter.com/share" class="twitter-share-button">Tweet</a>
-		</div>
-	</header>
-	<div id="main">
-
-		<div id="warningCanvas">Your browser does not seem to support the Canvas element. Please upgrade to a modern browser:<br/>
-		For instance <a href="http://www.mozilla.org/en-US/firefox/fx/">Firefox</a> or <a href="https://www.google.com/chrome">Chrome</a>
-		</div>
-		<div id="warningFileApi">Your browser does not seem to support the File API needed for drag&drop. Please upgrade to a modern browser:<br/>
-		For instance <a href="http://www.mozilla.org/en-US/firefox/fx/">Firefox</a> or <a href="https://www.google.com/chrome">Chrome</a>
-		</div>
-		<div id="dropArea">drag &amp; drop an image from your desktop to this page</div>
-		<div id="uploadForm"><span>Your browser does not seem to support the File API, however you can upload a image by inputting a address to a image:</span><input id="uploadURL" type="text" value="http://blur.fi/images/logo.png" autocomplete="off"/><a href="#">Upload</a></div>
-        <canvas id="uploadedPic"></canvas>
-        <div id="savedImage"></div>
-        <a href="#" id="blurAll">BLUR IMAGE</a>
-    </div>
-    <footer>
-    <div id="settings">
-		<h3>settings</h3>
-		<div class="settingsBlock">
-			<h5 id="displayPixelSize">pixelsize: 15</h5>
-			<div id="slider"></div>
-			<p>
-				<input type="checkbox" class="check" id="mouseBlur"> mouseblur
-			</p>
-			<p>
-			<input type="checkbox" class="check" id="scaleImage" checked > scale to fit
-			</p>
-		</div>
-		<div class="settingsBlock">
-		<h5 id="displayBlurStyle">blurstyle: Default</h5>
-		<ul id="styleSettings">
-		<li>
-			<a href="#" title="Default" class="styleSelector active" data-blur-style="default" id="blurStyleDefault"></a>
-		</li>
-		<li>
-			<a href="#" title="Circle" class="styleSelector" data-blur-style="circular" id="blurStyleCircular"></a>
-		</li>
-		<li>
-			<a href="#" title="Circle, color background" class="styleSelector" data-blur-style="circularNegative" id="blurStyleCircularNegative"></a>
-		</li>
-		<li>
-			<a href="#" title="Ascii" class="styleSelector" data-blur-style="ascii" id="blurAscii"></a>
-		</li>
-		<li>
-			<a href="#" title="Rastered" class="styleSelector" data-blur-style="rastered" id="blurStyleRastered"></a>
-		</li>
-		<li>
-			<a href="#" title="Desaturated circle" class="styleSelector" data-blur-style="desaturatedCircle" id="blurStyleDesaturatedCircle"></a>
-		</li>
-		<li>
-			<a href="#" title="Lines" class="styleSelector" data-blur-style="lineArt" id="blurStyleLineArt"></a>
-		</li>
-		<li>
-			<a href="#" title="Cross stich" class="styleSelector" data-blur-style="crossStich" id="blurStyleCrossStich"></a>
-		</li>
-		<li>
-			<a href="#" title="Gradient" class="styleSelector" data-blur-style="gradient" id="blurStyleGradient"></a>
-		</li>
-	 	</ul>
-	 </div>
-		<div class="settingsBlock">
-			<h5>save</h5>
-			<a href="#" class="action" id="saveImage">save image</a>
-	 		<a href="#" class="action" id="saveToimgurl">save to imgurl</a>	 		
-	 	</div>
-	 	<div class="settingsBlock">
-	 		<h5>reset</h5>
-	 		<a href="#" class="action" id="revertImage">revert image</a>
-	 		<a href="#" class="action" id="reset">reset to default state</a>
-	 	</div>
-	 	<div class="clearfix"></div>
-	</div>
-</footer>
-
-<script type="text/javascript" src="js/libs/jquery-ui-1.8.17.custom.min.js"></script>
-<script type="text/javascript" src="dist/blur_fi.min.js"></script>
-<script type="text/javascript" >!function(d,s,id){var js,fjs=d.getElementsByTagName(s)[0];if(!d.getElementById(id)){js=d.createElement(s);js.id=id;js.src="//platform.twitter.com/widgets.js";fjs.parentNode.insertBefore(js,fjs);}}(document,"script","twitter-wjs");</script>
-</body>
+	<script src="./src/index.ts"></script>
+	<body></body>
 </html>
