--- js/blur-min.js
@@ -1,1 +1,0 @@
-if(!("bind" in Function.prototype)){Function.prototype.bind=function(a){var c=this;var b=Array.prototype.slice.call(arguments,1);return function(){return c.apply(a,b.length===0?arguments:arguments.length===0?b:b.concat(Array.prototype.slice.call(arguments,0)))}}}function Blur(a){this.pixelSize=a.pixelSize;this.uploadedImage=null;this.imageData=null;this.blurStyle=null;this.mouseBlur=false;this.uiInited=false;this.optimized=true;this.revertOnBlur=true;this.pointRegister=[];this.updatePoint=false;this.scaleImageToFit=true;this.pixelCollection=[];this.dropZone=document.getElementById(a.dropAreaId);this.targetCanvas=document.getElementById("uploadedPic");this.canvasSupport=null;this.fileReaderSupport=null;this.blurImageW=null;this.blurImageH=null;this.imgurlAPIkey="2b1e392999deaf5f7cac09a738aa3ae4";this.asciiCharMap=" .:-=+*#%@";this.asciiCharMapArr=this.asciiCharMap.split("");this.asciiCharMapArr.reverse();if(this.targetCanvas.getContext){this.canvasSupport=true;this.context=this.targetCanvas.getContext("2d");this.init()}else{this.canvasSupport=false;document.getElementById("warningCanvas").style.display="block";this.dropZone.style.display="none"}if(window.File&&window.FileReader&&window.FileList&&window.Blob){this.fileReaderSupport=true}else{this.fileReaderSupport=false;document.getElementById("warningFileApi").style.display="block";this.dropZone.style.display="none";if(this.canvasSupport){}}}Blur.prototype.init=function(){this.dropZone.addEventListener("dragover",this.onDragHandler,false);this.dropZone.addEventListener("dragleave",this.onDragHandler,false);this.dropZone.addEventListener("drop",this.onDrop.bind(this),false)};Blur.prototype.initUI=function(){document.getElementById("saveToimgurl").addEventListener("click",this.uploadToimgurl.bind(this),false);document.getElementById("saveImage").addEventListener("click",this.saveImage.bind(this),false);document.getElementById("blurAll").addEventListener("click",this.blurAll.bind(this),false);document.getElementById("revertImage").addEventListener("click",this.revertToOriginal.bind(this),false);document.getElementById("reset").addEventListener("click",this.reset.bind(this),false);document.getElementById("blurAll").style.display="block";document.getElementById("settings").style.display="block";this.uiInited=true};Blur.prototype.reset=function(a){a.preventDefault();this.context.clearRect(0,0,this.blurImageW,this.blurImageH);this.mouseBlur=false;this.targetCanvas.style.display="none";document.getElementById("blurAll").style.display="none";document.getElementById("mouseBlur").checked=false;document.getElementById("scaleImage").checked=false;if(this.fileReaderSupport){this.dropZone.style.display="block"}else{document.getElementById("uploadForm").style.display="block"}setPlacement()};Blur.prototype.uploadToimgurl=function(d){d.preventDefault();var c=this;try{var b=this.targetCanvas.toDataURL("image/jpeg",0.9).split(",")[1]}catch(d){var b=this.targetCanvas.toDataURL().split(",")[1]}var a=window.open();a.document.write("Uploading...");$.ajax({url:"http://api.imgur.com/2/upload.json",type:"POST",data:{type:"base64",key:c.imgurlAPIkey,name:"blur.jpg",title:"Image created by http://www.blur.fi",caption:"Pixelate your images in your broser",image:b},dataType:"json"}).success(function(e){a.location.href=e.upload["links"]["imgur_page"];$.ajax({url:"http://blur.fi/php/bridge.php?url=https://api.mongolab.com/api/1/databases/blur_fi_images/collections/gallery",data:JSON.stringify({data:e.upload}),type:"POST",contentType:"application/json",success:function(f){}})}).error(function(){alert("Could not reach api.imgur.com. Sorry :(");a.close()})};Blur.prototype.setPixelSize=function(a){this.pixelSize=a;this.pointRegister=[]};Blur.prototype.setBlurStyle=function(a){this.blurStyle=a;this.pointRegister=[]};Blur.prototype.scaleToFit=function(a){this.scaleImageToFit=a;this.imgLoaded(this.uploadedImage)};Blur.prototype.blurWithMouse=function(a){this.mouseBlur=a;if(this.mouseBlur){this.targetCanvas.classList.add("mouseblur")}else{this.targetCanvas.classList.remove("mouseblur")}this.targetCanvas.addEventListener("mousedown",this.mousepresshandler.bind(this),false);window.addEventListener("mouseup",this.mousepresshandler.bind(this),false);this.targetCanvas.addEventListener("mousemove",this.mousemovehandler.bind(this),false)};Blur.prototype.onDragHandler=function(a){a.preventDefault();if(a.type==="dragover"){document.body.classList.add("active")}else{document.body.classList.remove("active")}};Blur.prototype.onDrop=function(h){var d=this;h.preventDefault();var c=h.dataTransfer.files;for(var b=0;b<c.length;b++){var a=new FileReader();var g=c[b];a.onload=(function(e){return function(i){var f=new Image();f.onload=function(){d.imgLoaded(f)};f.src=i.target.result}})(g);a.readAsDataURL(g)}};Blur.prototype.revertToOriginal=function(a){if(a){a.preventDefault()}this.context.clearRect(0,0,this.blurImageW,this.blurImageH);this.context.drawImage(this.uploadedImage,0,0,this.blurImageW,this.blurImageH)};Blur.prototype.mousepresshandler=function(a){a.preventDefault();this.targetCanvas.style.cursor=a.type==="mousedown"?"pointer":"default";this.updatePoint=a.type==="mousedown"?true:false};Blur.prototype.mousemovehandler=function(a){if(this.updatePoint){this.blurCoords(a.pageX,a.pageY)}if(this.pixelCollection.length>0){this.blurBlock()}if((!this.updatePoint)&&(this.pixelCollection.length===0)){this.targetCanvas.removeEventListener("mousemove",this.mousemovehandler.bind(this),false)}};Blur.prototype.imgLoaded=function(b){document.body.classList.remove("active");this.dropZone.style.display="none";document.getElementById("uploadForm").style.display="none";this.uploadedImage=b;this.blurImageW=this.uploadedImage.width;this.blurImageH=this.uploadedImage.height;if(this.scaleImageToFit){if((this.blurImageW>window.innerWidth)||(this.blurImageH>window.innerHeight)){var a=Math.min(window.innerWidth/this.blurImageW,window.innerHeight/this.blurImageH);this.blurImageW=a*this.blurImageW;this.blurImageH=a*this.blurImageH;document.getElementById("scaleImage").disabled=false}else{document.getElementById("scaleImage").disabled=true}}this.targetCanvas.style.width=this.targetCanvas.width=this.blurImageW;this.targetCanvas.style.height=this.targetCanvas.height=this.blurImageH;this.context.drawImage(b,0,0,this.blurImageW,this.blurImageH);this.targetCanvas.style.display="block";setPlacement();if(!this.uiInited){this.initUI()}else{document.getElementById("blurAll").style.display="block"}};Blur.prototype.saveImage=function(c){c.preventDefault();var a=this.context.canvas.toDataURL("image/png");var b=175;var d=Math.floor(b*(this.targetCanvas.height/this.targetCanvas.width));$("#savedImage").html("<h5>RIGHT CLICK ON WINDOWS OR CTRL-CLICK ON MAC TO SAVE IMAGE. ON CHROME YOU CAN ALSO DRAG YOUR IMAGE TO YOUR DESKTOP</h5><br/><img width="+b+" height="+d+" title='Save image' draggable='true' src='"+a+"'/>");$("#savedImage").slideDown(200)};Blur.prototype.blurAll=function(x){x.preventDefault();if(this.revertOnBlur){this.revertToOriginal()}var i=this;var b,d,s,y,v,o,g,k,i,t,r;d=new Date();t=i.pixelSize;var z=i.targetCanvas.height;var m=i.targetCanvas.width;var f={};f.w=f.h=t;var c=this.context;var b=i.pixelSize*i.pixelSize*4;if(i.imageData!==null){s=i.imageData}else{var y=i.context.getImageData(0,0,m,z);s=i.imageData=y.data}var p=m/t;var n=z/t;var q=[];var a=null;for(g=0;g<p;g++){for(k=0;k<n;k++){f.h=Math.min(z-k*t,t);var j=greenCh=blueCh=0;for(v=0;v<f.h;v++){f.w=Math.min(m-g*t,t);for(o=0;o<f.w;o++){a=((t*(k)+v)*m+(t*(g)+o))*4;j+=s[a];greenCh+=s[a+1];blueCh+=s[a+2]}}b=f.h*f.w*4;q[q.length]={r:j,g:greenCh,b:blueCh,y:t*k,x:t*g,pixelW:f.w,pixelH:f.h,loopAmount:b}}}this.drawPixels(q);var u=new Date();r=d.getTime();var l=u.getTime()-r;if(location.search==="?debug"){alert(l)}_gaq.push(["_trackEvent","BlurRuntime",l+" ms"]);_gaq.push(["_trackEvent","BlurImageSize",m+"x"+z]);_gaq.push(["_trackEvent","BlurPixelSize",t+" px"]);_gaq.push(["_trackEvent","BlurPixelStyle","pixelStyle",this.blurStyle]);return false};Blur.prototype.drawPixels=function(a){this.blurFillOnePass(a)};Blur.prototype.setPixel=function(k,d,j,i,h,c,e){var f=(d+j*k.width)*4;k.data[f+0]=i;k.data[f+1]=h;k.data[f+2]=c;k.data[f+3]=e};Blur.prototype.blurFillOnePass=function(c){var b=0;var a=c.length;switch(this.blurStyle){case"ascii":this.context.fillStyle="rgba(255,255,255, 1)";this.context.fillRect(0,0,this.targetCanvas.width,this.targetCanvas.height);this.context.fillStyle="rgba(0,0,0,1)";this.context.font=this.pixelSize+"px Courier";break;case"crossStich":this.context.fillStyle="rgba(0,0,0, 1)";this.context.fillRect(0,0,this.targetCanvas.width,this.targetCanvas.height);break;case"circular":this.context.fillStyle="rgba(20,20,20, 1)";this.context.fillRect(0,0,this.targetCanvas.width,this.targetCanvas.height);break;case"rastered":this.context.fillStyle="rgba(255,255,255, 1)";this.context.fillRect(0,0,this.targetCanvas.width,this.targetCanvas.height);break}for(b=0;b<a;b++){this.blurFill(c[b],true)}};Blur.prototype.blurFill=function(e,p){var m=e.r;var v=e.g;var A=e.b;var i=e.x;var h=e.y;var f=e.pixelW;var l=e.pixelH;var a=e.loopAmount;var q=this.getRgbValue(m,v,A,a);switch(this.blurStyle){case"softblur":this.setPixel(this.collPixels,i/this.pixelSize,h/this.pixelSize,m,v,A,255);break;case"circular":if(!p){this.context.fillStyle="rgba(20,20,20, 1)";this.context.fillRect(i,h,f,l)}this.context.fillStyle="rgba("+q.r+", "+q.g+", "+q.b+", 0.7)";this.context.beginPath();this.context.arc(i+this.pixelSize/2,h+this.pixelSize/2,this.pixelSize/2,0,Math.PI*2,true);this.context.closePath();this.context.fill();break;case"circularNegative":this.context.fillStyle="rgba("+Math.round(m/(a>>2))+", "+Math.round(v/(a>>2))+", "+Math.round(A/(a>>2))+", 1)";this.context.fillRect(i,h,f,l);this.context.fillStyle="rgba(20,20,20, 1)";this.context.beginPath();this.context.arc(i+this.pixelSize/2,h+this.pixelSize/2,this.pixelSize/4,0,Math.PI*2,true);this.context.closePath();this.context.fill();break;case"invertedBackdrop":var z=this.getRgbValue(m,v,A,a,true);this.context.fillStyle="rgba("+q.r+", "+q.g+", "+q.b+", 1)";this.context.fillRect(i,h,f,l);this.context.fillStyle="rgba("+z.r+", "+z.g+", "+z.b+", 1)";this.context.beginPath();this.context.arc(i+this.pixelSize/2,h+this.pixelSize/2,this.pixelSize/4,0,Math.PI*2,true);this.context.closePath();this.context.fill();break;case"invert":var z=this.getRgbValue(m,v,A,a,true);this.context.fillStyle="rgba("+z.r+", "+z.g+", "+z.b+", 1)";this.context.fillRect(i,h,f,l);break;case"desaturated":q=this.getRgbValue(m,v,A,a,false,true);this.context.fillStyle="rgba("+q.r+", "+q.g+", "+q.b+", 1)";this.context.fillRect(i,h,f,l);break;case"crossStich":if(!p){this.context.fillStyle="rgba(0,0,0, 1)";this.context.fillRect(i,h,f,l)}this.context.beginPath();this.context.fillStyle="rgba("+q.r+", "+q.g+", "+q.b+", 1)";this.context.strokeStyle="rgba("+q.r+", "+q.g+", "+q.b+", 1)";this.context.lineWidth=Math.round(this.pixelSize/4);this.context.moveTo(i,h);this.context.lineTo(i+this.pixelSize,h+this.pixelSize);this.context.moveTo(i+this.pixelSize,h);this.context.lineTo(i,h+this.pixelSize);this.context.lineCap="square";this.context.stroke();this.context.closePath();break;case"gradient":var w=Math.round(i+this.pixelSize/2);var d=Math.round(h+this.pixelSize/2);var t=this.pixelSize/5;var u=w;var c=d;var s=this.pixelSize/1.3;var o=this.context.createRadialGradient(w,d,t,u,c,s);o.addColorStop(0,"rgb("+Math.round(m/(a>>2))+","+Math.round(v/(a>>2))+", "+Math.round(A/(a>>2))+")");o.addColorStop(0.7,"rgb(0, 0, 0)");this.context.fillStyle=o;this.context.fillRect(i,h,f,l);break;case"lineArt":this.context.fillStyle="rgba(0,0,0, 1)";this.context.fillRect(i,h,f,l);this.context.beginPath();q=this.getRgbValue(m,v,A,a,false,false);this.context.fillStyle="rgba("+q.r+", "+q.g+", "+q.b+", 1)";this.context.strokeStyle="rgba("+q.r+", "+q.g+", "+q.b+", 1)";this.context.lineWidth=Math.round(this.pixelSize/4);this.context.moveTo(i+this.pixelSize,h);this.context.lineTo(i+this.pixelSize,h+this.pixelSize);this.context.lineCap="square";this.context.stroke();this.context.closePath();break;case"desaturatedCircle":this.context.fillStyle="rgba(255,255,255, 1)";this.context.fillRect(i,h,f,l);q=this.getRgbValue(m,v,A,a,false,true);this.context.fillStyle="rgba("+q.r+", "+q.g+", "+q.b+", 1)";this.context.beginPath();this.context.arc(i+this.pixelSize/2,h+this.pixelSize/2,this.pixelSize/2,0,Math.PI*2,true);this.context.closePath();this.context.fill();break;case"ascii":if(!p){this.context.fillStyle="rgba(255,255,255, 1)";this.context.fillRect(i,h,f,l);this.context.fillStyle="rgba(0,0,0,1)";this.context.font=this.pixelSize+"px Courier"}q=this.getRgbValue(m,v,A,a,false,true);var j=Math.max(0,Math.min(this.asciiCharMapArr.length,Math.floor((q.r-1)/(255/this.asciiCharMapArr.length))));this.context.fillText(this.asciiCharMapArr[j],i,h);break;case"bw":q=this.getRgbValue(m,v,A,a,false,true);var n=q.r<128?0:255;this.context.fillStyle="rgba("+n+", "+n+", "+n+", 1)";this.context.fillRect(i,h,f,l);break;case"rastered":if(!p){this.context.fillStyle="rgba(255,255,255, 1)";this.context.fillRect(i,h,f,l)}q=this.getRgbValue(m,v,A,a,false,true);this.context.fillStyle="rgba(0,0,0,1)";this.context.beginPath();var k=q.r/255;this.context.arc(i+this.pixelSize/2,h+this.pixelSize/2,Math.round(this.pixelSize/2-((this.pixelSize/2)*k)),0,Math.PI*2,true);this.context.closePath();this.context.fill();break;default:this.context.fillStyle="rgba("+q.r+", "+q.g+", "+q.b+", 1)";this.context.fillRect(i,h,f,l);break}};Blur.prototype.getRgbValue=function(i,h,a,d,c,f){var e={r:Math.round(i/(d>>2)),g:Math.round(h/(d>>2)),b:Math.round(a/(d>>2))};if(c){e.r=Math.abs(e.r-255);e.g=Math.abs(e.g-255);e.b=Math.abs(e.b-255)}if(f){e.r=e.g=e.b=Math.round(e.r*0.3+e.g*0.59+e.b*0.11)}return e};Blur.prototype.blurBlock=function(){if(this.pixelCollection.length>0){var h=this.pixelCollection.shift();var f,d,j,c,a;var b={};b.w=b.h=null;b.h=h.y+this.pixelSize<this.targetCanvas.height?this.pixelSize:this.targetCanvas.height-h.y;b.w=h.x+this.pixelSize<this.targetCanvas.width?this.pixelSize:this.targetCanvas.width-h.x;var g=b.w*b.h*4;if(g===0){return false}a=this.context.getImageData(Math.min(h.x,this.targetCanvas.width),Math.min(h.y,this.targetCanvas.height),b.w,b.h);c=a.data;f=d=j=0;var e=0;while(e<g){f+=c[e];j+=c[++e];d+=c[++e];e+=2}this.blurFill({r:f,g:j,b:d,y:h.y,x:h.x,pixelW:b.w,pixelH:b.h,loopAmount:g})}};Blur.prototype.blurCoords=function(d,c){if(this.mouseBlur){var b=this.snapCoord(d-this.targetCanvas.offsetLeft);var a=this.snapCoord(c-this.targetCanvas.offsetTop);if(this.pointRegister[b+"x"+a]===undefined){this.pixelCollection.push({x:b,y:a});this.pointRegister[b+"x"+a]=true}}};Blur.prototype.snapCoord=function(b){var a=0;a=b<this.pixelSize?0:(Math.floor(b/this.pixelSize))*this.pixelSize;return a};var blur=new Blur({pixelSize:15,dropAreaId:"dropArea"});$(".sizeSelector").click(function(a){a.preventDefault();blur.setPixelSize($(a.currentTarget).data("pixel-size"))});$(".styleSelector").click(function(a){a.preventDefault();$(".styleSelector").removeClass("active");$(a.currentTarget).addClass("active");$("#displayBlurStyle").text("blurstyle: "+$(a.currentTarget).attr("title"));blur.setBlurStyle($(a.currentTarget).data("blur-style"))});function setPlacement(){$("footer").css({width:$("#settings").width()});if($(blur.targetCanvas).is(":hidden")){var b=$(window).height();b-=$("header").outerHeight()+10;b-=$("footer").height();$("#main").height(b);var a=Math.floor((($(window).height()-$("header").outerHeight()-$("footer").height())-$("#dropArea").height())/2);$("#dropArea").addClass("active");$("#dropArea").css({"padding-top":a+"px"})}else{$("#main").css({height:""})}}$(window).resize(function(){setPlacement()});$(document).ready(function(){setPlacement()});$("#mouseBlur, #scaleImage").change(function(a){a.preventDefault();if($(a.currentTarget).attr("id")==="mouseBlur"){blur.blurWithMouse($(a.currentTarget).is(":checked"))}else{blur.scaleToFit($(a.currentTarget).is(":checked"))}});$("#slider").slider({max:50,min:3,value:10,step:1,slide:function(a,b){$("#displayPixelSize").text("Pixelsize: "+b.value)},stop:function(a,b){blur.setPixelSize(b.value)}});$("#uploadForm a").click(function(b){b.preventDefault();var a=$("#uploadURL").val();$.getImageData({url:a,success:function(c){blur.imgLoaded(c)},error:function(){$("#uploadURL").val("Doh, something went wrong, try again later.")}})});if((navigator.userAgent.match(/iPhone/i))||(navigator.userAgent.match(/iPod/i))||(navigator.userAgent.match(/iPad/i))){$("#mouseBlur").hide()};
