--- js/ui-min.js
@@ -1,1 +1,0 @@
-$(".sizeSelector").click(function(a){a.preventDefault();blur.setPixelSize($(a.currentTarget).data("pixel-size"))});$(".styleSelector").click(function(a){a.preventDefault();$(".styleSelector").removeClass("active");$(a.currentTarget).addClass("active");$("#displayBlurStyle").text("Blur style: "+$(a.currentTarget).attr("title"));blur.setBlurStyle($(a.currentTarget).data("blur-style"))});$("#settingsVisibility").click(function(b){b.preventDefault();$("#settings").slideToggle(150).toggleClass("closed");var a=$("#settings").hasClass("closed")?"Show":"Hide";a+=" settings";$(b.currentTarget).text(a)});$("#mouseBlur").click(function(b){b.preventDefault();$(b.currentTarget).toggleClass("active");blur.blurWithMouse($(b.currentTarget).hasClass("active"));var a="Mouseblur ";a+=$(b.currentTarget).hasClass("active")?"on":"off";$(b.currentTarget).text(a)});$("#iOsPixelSelector > ul > li").click(function(b){b.preventDefault();var a=$(b.currentTarget).data("pixelsize");blur.setPixelSize(a);$("#displayPixelSize").text("Pixelsize: "+a)});$("#slider").slider({max:50,min:4,value:15,step:1,slide:function(a,b){$("#displayPixelSize").text("Pixelsize: "+b.value)},stop:function(a,b){blur.setPixelSize(b.value)}});$("#uploadForm a").click(function(b){b.preventDefault();var a=$("#uploadURL").val();$.getImageData({url:a,success:function(c){blur.imgLoaded(c)},error:function(){$("#uploadURL").val("Doh, something went wrong, try again later.")}})});if((navigator.userAgent.match(/iPhone/i))||(navigator.userAgent.match(/iPod/i))||(navigator.userAgent.match(/iPad/i))){$("#iOsPixelSelector").show();$("#slider,#mouseBlur").hide()};
