+++ src/index.ts
@@ -1,0 +1,1 @@
+console.log('hello world')
